requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
antlr4-python3-runtime==4.9.3
    # via omegaconf
audioread==3.0.1
    # via librosa
certifi==2025.8.3
    # via requests
cffi==2.0.0
    # via soundfile
charset-normalizer==3.4.3
    # via requests
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
decorator==5.2.1
    # via librosa
filelock==3.19.1
    # via torch
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.60.0
    # via matplotlib
fsspec==2025.9.0
    # via torch
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via requests
jinja2==3.1.6
    # via torch
joblib==1.5.2
    # via
    #   librosa
    #   scikit-learn
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via librosa
librosa==0.11.0
    # via -r requirements.in
llvmlite==0.45.0
    # via numba
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.6
    # via
    #   -r requirements.in
    #   seaborn
mpmath==1.3.0
    # via sympy
msgpack==1.1.1
    # via librosa
networkx==3.5
    # via torch
numba==0.62.0
    # via librosa
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   librosa
    #   matplotlib
    #   numba
    #   onnxruntime
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   soundfile
    #   soxr
omegaconf==2.3.0
    # via silero
onnxruntime==1.22.1
    # via -r requirements.in
packaging==25.0
    # via
    #   lazy-loader
    #   matplotlib
    #   onnxruntime
    #   pooch
pandas==2.3.2
    # via
    #   -r requirements.in
    #   seaborn
pillow==11.3.0
    # via matplotlib
platformdirs==4.4.0
    # via pooch
pooch==1.8.2
    # via librosa
protobuf==6.32.1
    # via onnxruntime
pycparser==2.23
    # via cffi
pydub==0.25.1
    # via -r requirements.in
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   omegaconf
requests==2.32.5
    # via
    #   -r requirements.in
    #   pooch
scikit-learn==1.7.2
    # via librosa
scipy==1.16.2
    # via
    #   librosa
    #   scikit-learn
seaborn==0.13.2
    # via -r requirements.in
silero==0.4.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
soundfile==0.13.1
    # via librosa
soxr==1.0.0
    # via librosa
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
textgrid==1.6.1
    # via -r requirements.in
threadpoolctl==3.6.0
    # via scikit-learn
torch==2.2.2
    # via
    #   -r requirements.in
    #   silero
    #   torchaudio
torchaudio==2.2.2
    # via
    #   -r requirements.in
    #   silero
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   librosa
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
